Added back coverage (broken), slight refactoring of server schema files
