+++ .travis.yml
@@ -6,6 +6,10 @@ services:
 before_install:
   - sleep 15
   - mongo mydb_test --eval 'db.addUser("travis", "test");'
-  - npm install -g grunt-cli
+  - npm install -g gulp-cli
 install:
   - npm install
+before_script:
+  - gulp
+after_script:
+  - npm run coverage
