+++ package.json
@@ -4,7 +4,8 @@
   "description": "Udia - Commits for the sake of commits",
   "scripts": {
     "start": "node ./bin/www",
-    "test": "mocha --reporter spec --compilers ts:ts-node/register test/**/*.test.ts"
+    "test": "mocha --reporter spec --compilers ts:ts-node/register test/**/*.test.ts",
+    "coverage": "istanbul cover -e .ts mocha -- --reporter spec --compilers ts:ts-node/register"
   },
   "repository": {
     "type": "git",
@@ -31,10 +32,10 @@
     "method-override": "^2.3.7",
     "mongoose": "^4.7.2",
     "morgan": "^1.7.0",
-    "reflect-metadata": "^0.1.8",
-    "rxjs": "^5.0.0-rc.4",
+    "reflect-metadata": "0.1.8",
+    "rxjs": "5.0.0-rc.4",
     "systemjs": "0.19.40",
-    "zone.js": "^0.7.2"
+    "zone.js": "0.7.2"
   },
   "devDependencies": {
     "@types/bluebird": "^3.0.36",
@@ -57,6 +58,7 @@
     "del": "^2.2.2",
     "gulp": "^3.9.1",
     "gulp-typescript": "^3.1.3",
+    "istanbul": "^1.1.0-alpha.1",
     "mocha": "^3.2.0",
     "ts-node": "^1.7.1",
     "typescript": "^2.1.4"
