+++ src/client/index.html
@@ -4,6 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <title>UDIA</title>
+  <link rel="icon" href="/favicon.ico" />
   <base href="/">
 
   <!-- ngdeps and config are populated by the gulp build process and will exist in /dist/client -->
