+++ src/server/api/thing/thing.schema.ts
@@ -3,11 +3,14 @@
  */
 import * as mongoose from "mongoose";
 
+// To reference this interface:
+// import {Thing} from "./thing.schema";
 export interface Thing {
   message: string;
   createdAt: Date;
 }
-
+// To reference this mongoose.Schema:
+// import thingSchema from "./thing.schema";
 let schema = new mongoose.Schema({
   message: {
     type: String, required: true, trim: true
